Move climatological processing to separate module
